requirements.txt
pandas
#psycopg2-binary>=2.8.6
psycopg2-binary>=2.9.9
apache-airflow-providers-postgres==3.0.0
apache-airflow-providers-google==6.0.0
apache-airflow[google]


